requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
fsspec==2025.9.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
polars==1.34.0
polars-runtime-32==1.34.0
psutil==7.1.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scipy==1.15.3
six==1.17.0
sympy==1.14.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
typing_extensions==4.15.0
ultralytics==8.3.218
ultralytics-thop==2.0.17
urllib3==2.5.0
Werkzeug==3.1.3
